Prep my day working with activity model.
